requirements.txt
psycopg2-binary       # Работа с PostgreSQL
sqlalchemy            # ORM для удобной работы с БД
appium-python-client  # Управление Android-эмулятором через Appium
requests              # Для работы с API (например, sms-activate.org)
python-dotenv
imapclient
PyQt5
